多个源读书路径 src

+++ thiefFish/GetBook1.py
@@ -7,6 +7,7 @@ from fake_useragent import UserAgent
 import sys
 import random
 import json
+import urllib.parse 
 
 proxies = [{}]
 #proxies = [{'http': 'http://59.55.161.88:3256', 'https': 'https://59.55.161.88:3256'},   {'http': 'http://103.37.141.69:80', 'https': 'https://103.37.141.69:80'},   {'http': 'http://124.205.153.36:80', 'https': 'https://124.205.153.36:80'},   {'http': 'http://60.191.11.241:3256', 'https': 'https://60.191.11.241:3256'},   {'http': 'http://175.7.199.229:6969', 'https': 'https://175.7.199.229:6969'},   {'http': 'http://27.191.60.36:8888', 'https': 'https://27.191.60.36:8888'},   {'http': 'http://121.4.36.93:8888', 'https': 'https://121.4.36.93:8888'}]
@@ -15,8 +16,9 @@ def href_Byname(book_name):
     通过传入的书号bookid，获取此书的所有章节目录
     :param book_id:
     :return: 章节目录及章节地址
-    '''
-    url = 'http://www.biquw.la/modules/article/search.php?action=login&searchkey={}'.format(book_name)
+    ''' 
+    encoded_book_name = urllib.parse.quote(book_name[1].encode('gbk'))  
+    url = 'https://www.biquge5.la/modules/article/search.php?searchkey={}&searchtype={}'.format(encoded_book_name,'articlename')
     headers = {
         'User-Agent':random_agent()}
         #'User-Agent': UserAgent().random}
@@ -24,7 +26,7 @@ def href_Byname(book_name):
     response = requests.get(url, headers,proxies=random.choice(proxies))
     response.encoding = response.apparent_encoding
     response = BeautifulSoup(response.text, 'lxml')
-    href = response.find('li').find('a').get('href')
+    href = response.find('link').get('href')
     return href
 
 def book_page_list(book_id):
@@ -33,7 +35,7 @@ def book_page_list(book_id):
     :param book_id:
     :return: 章节目录及章节地址
     '''
-    url = 'http://www.biquw.com/book/{}/'.format(book_id)
+    url = 'https://www.biquge5.la/book/{}/'.format(book_id)
     headers = {
         'User-Agent':random_agent()}
         #'User-Agent': UserAgent().random}
@@ -41,7 +43,11 @@ def book_page_list(book_id):
     response = requests.get(url, headers,proxies=random.choice(proxies))
     response.encoding = response.apparent_encoding
     response = BeautifulSoup(response.text, 'lxml')
-    booklist = response.find('div', class_='book_list').find_all('a')
+    # 找到特定的 class  
+    start_div = response.find(id='list').find(class_='clear')  
+
+    # 在 start_div 下查找所有 <a> 标签  
+    booklist = start_div.find_all_next('a') 
     return booklist
 
 
@@ -52,12 +58,13 @@ def book_page_text(bookid, booklist):
     :param booklist:
     :return:None
     '''
-    try:
+    if os.path.exists('./txtPaths'):  
         os.remove('./txtPaths')
+    try:
         for book_page in booklist:
             page_name = book_page.text.replace('*', '')
             page_id = book_page['href']
-            url = 'http://www.biquw.com/book/{}/{}'.format(bookid,page_id)
+            url = 'https://www.biquge5.la/book/{}/{}'.format(bookid,page_id)
             headers = {
                 'User-Agent':random_agent()}
                 #'User-Agent': UserAgent().random}
@@ -65,9 +72,10 @@ def book_page_text(bookid, booklist):
             response_book = requests.get(url, headers,proxies=random.choice(proxies))
             response_book.encoding = response_book.apparent_encoding
             response_book = BeautifulSoup(response_book.text, 'lxml')
-            book_content = response_book.find('div', id="htmlContent")
-            with open("./books/{}/{}.txt".format(bookid,page_name), 'a') as f:
-                f.write(book_content.text.replace('\xa0', ''))
+            book_content = response_book.find('div', id="content")
+            with open("./books/{}/{}.txt".format(bookid,page_name), 'a') as f:  
+                book_content_wrap = wrap_string(book_content.text.replace('\xa0', '').replace('\ufffd', ''),58)
+                f.write(book_content_wrap)
                 print("当前下载章节：{}".format(page_name))
             with open("./txtPaths", 'a') as f:
                 f.write("./books/{}/{}.txt".format(bookid,page_name)+"\n")
@@ -81,12 +89,16 @@ def random_agent():
         data = json.load(file)  
         return random.choice(data) 
 
+def wrap_string(string, line_length):  
+    # 使用列表推导式和 join 方法  
+    wrapped_lines = [string[i:i + line_length] for i in range(0, len(string), line_length)]  
+    return '\n'.join(wrapped_lines)  
+
 def getbook_main(bookname):
     href = href_Byname(bookname)
     time.sleep(1)
     bookid= href.split('/')[-2]
     
-    print(bookid)
     # 如果书号对应的目录不存在，则新建目录，用于存放章节内容
     if not os.path.isdir('./books/{}'.format(bookid)):
         os.mkdir('./books/{}'.format(bookid))
